Stop creating lists on ticket and device placement
Updated the router so that when the device or ticket placement loads, it
does not create a new list automatically.
Now, any time any of the placements load without a list ID, they create 
an unsaved list object, which will only get save once an update is made 
to the list.
Simplified the router, so that the SW ticket and device info are 
gathered at the same time before the client side redirect.
Fixed the height of the top nav on the device and ticket placements to 
prevent jumping when the nav gets its first list.
Changed the default list title to simply ‘My New List’.
No longer display the new list, copy list, or delete list buttons on a 
new list that has yet to be saved.
Now send up the entire list object when any field is updated, just to 
keep it simple with a single code path.

Fixes #1.